backend/requirements.txt
uvicorn==0.32.0
fastapi==0.115.3
google-generativeai==0.8.3
python-dotenv==1.0.1
supabase==2.9.1